SEQUOIADBMAINSTREAM-2882 整形值乘法溢出问题
git-svn-id: http://192.168.20.11/sequoiadb/trunk@33163 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/dmsStorageDataCapped.hpp
@@ -421,7 +421,7 @@ namespace engine
       const dmsMBStatInfo *mbStatInfo = getMBStatInfo( context->mbID() ) ;
       SDB_ASSERT( mbStatInfo, "mbStatInfo should not be NULL" ) ;
 
-      return (( mbStatInfo->_totalDataPages << pageSizeSquareRoot()) + newSize)
+      return (((UINT64)mbStatInfo->_totalDataPages << pageSizeSquareRoot()) + newSize)
              > (UINT64)_options[context->mbID()]->_maxSize ;
    }
